# clients/java/api/openapi.yaml
openapi: 3.1.0
info:
  description: "Define, evaluate, and enforce data retention policies"
  title: DRPE — Data Retention Policy Engine
  version: 0.1.0
servers:
- description: Local DRPE API
  url: http://localhost:8000
paths:
...
  /api/v1/classify:
    post:
      operationId: classify_one_api_v1_classify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassificationRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassificationResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Classify One
      tags:
      - classify
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/classify/batch:
    post:
      operationId: classify_batch_api_v1_classify_batch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchClassificationRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ClassificationResponse"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Classify Batch
      tags:
      - classify
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/classify/dry-run:
    post:
      operationId: classify_dry_run_api_v1_classify_dry_run_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassificationRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassificationResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Classify Dry Run
      tags:
      - classify
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/evaluate:
    post:
      operationId: evaluate_one_api_v1_evaluate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Evaluate One
      tags:
      - evaluate
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/evaluate/batch:
    post:
      operationId: evaluate_batch_api_v1_evaluate_batch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchEvaluateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/EvaluationResponse"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Evaluate Batch
      tags:
      - evaluate
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/evaluate/dry-run:
    post:
      operationId: evaluate_dry_run_api_v1_evaluate_dry_run_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Evaluate Dry Run
      tags:
      - evaluate
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/enforce:
    post:
      operationId: trigger_enforce_api_v1_enforce_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnforceRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnforceResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Trigger Enforce
      tags:
      - enforce
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/enforce/jobs:
    get:
      operationId: list_jobs_api_v1_enforce_jobs_get
      parameters:
      - explode: true
        in: query
        name: status
        required: false
        schema:
          $ref: "#/components/schemas/JobStatus"
        style: form
      - explode: true
        in: query
        name: policy_id
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
        style: form
      - explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/EnforcementJob"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: List Jobs
      tags:
      - enforce
      x-accepts:
      - application/json
  /api/v1/enforce/jobs/{job_id}:
    get:
      operationId: get_job_api_v1_enforce_jobs__job_id__get
      parameters:
      - explode: false
        in: path
        name: job_id
        required: true
        schema:
          title: Job Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnforcementJob"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Job
      tags:
      - enforce
      x-accepts:
      - application/json
  /api/v1/dsar/access:
    post:
      operationId: submit_access_api_v1_dsar_access_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DsarSubmitRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DsarRequest"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Submit Access
      tags:
      - dsar
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/dsar/erasure:
    post:
      operationId: submit_erasure_api_v1_dsar_erasure_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DsarSubmitRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DsarRequest"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Submit Erasure
      tags:
      - dsar
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/v1/dsar/requests:
    get:
      operationId: list_requests_api_v1_dsar_requests_get
      parameters:
      - explode: true
        in: query
        name: type
        required: false
        schema:
          $ref: "#/components/schemas/DsarRequestType"
        style: form
      - explode: true
        in: query
        name: status
        required: false
        schema:
          $ref: "#/components/schemas/DsarRequestStatus"
        style: form
      - explode: true
        in: query
        name: subject_id
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - explode: true
        in: query
        name: policy_id
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
        style: form
      - explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/DsarRequest"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: List Requests
      tags:
      - dsar
      x-accepts:
      - application/json
  /api/v1/dsar/requests/{request_id}:
    get:
      operationId: get_request_api_v1_dsar_requests__request_id__get
      parameters:
      - explode: false
        in: path
        name: request_id
        required: true
        schema:
          title: Request Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DsarRequest"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get Request
      tags:
      - dsar
      x-accepts:
      - application/json
...
  /api/v1/jurisdictions:
    get:
      operationId: list_all_api_v1_jurisdictions_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
          description: Successful Response
      security:
      - BearerAuth: []
      summary: List All
      tags:
      - jurisdictions
      x-accepts:
      - application/json
  /api/v1/jurisdictions/{code}:
    get:
      operationId: get_one_api_v1_jurisdictions__code__get
      parameters:
      - explode: false
        in: path
        name: code
        required: true
        schema:
          title: Code
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                title: Response Get One Api V1 Jurisdictions  Code  Get
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Get One
      tags:
      - jurisdictions
      x-accepts:
      - application/json
...
components:
  schemas:
...
    BatchClassificationRequest:
      example:
        records:
        - record_id: record_id
          metadata:
            key: ""
          policy_id: policy_id
          jurisdiction: jurisdiction
          text_fields:
          - text_fields
          - text_fields
          data_type: data_type
          source: source
        - record_id: record_id
          metadata:
            key: ""
          policy_id: policy_id
          jurisdiction: jurisdiction
          text_fields:
          - text_fields
          - text_fields
          data_type: data_type
          source: source
      properties:
        records:
          items:
            $ref: "#/components/schemas/ClassificationRequest"
          type: array
      required:
      - records
      title: BatchClassificationRequest
    BatchEvaluateRequest:
      example:
        records:
        - record_id: record_id
          metadata:
            key: ""
          jurisdiction: jurisdiction
          data_type: data_type
          context:
            key: ""
          source: source
        - record_id: record_id
          metadata:
            key: ""
          jurisdiction: jurisdiction
          data_type: data_type
          context:
            key: ""
          source: source
      properties:
        records:
          items:
            $ref: "#/components/schemas/EvaluationRequest"
          type: array
      required:
      - records
      title: BatchEvaluateRequest
...
    ClassificationRequest:
      description: Request to classify a record against classification policies.
      example:
        record_id: record_id
        metadata:
          key: ""
        policy_id: policy_id
        jurisdiction: jurisdiction
        text_fields:
        - text_fields
        - text_fields
        data_type: data_type
        source: source
      properties:
        data_type:
          title: Data Type
          type: string
        record_id:
          title: Record Id
          type: string
        metadata:
          additionalProperties: true
          title: Metadata
          type: object
        source:
          nullable: true
          type: string
        jurisdiction:
          nullable: true
          type: string
        text_fields:
          items:
            type: string
          nullable: true
          type: array
        policy_id:
          nullable: true
          type: string
      required:
      - data_type
      - record_id
      title: ClassificationRequest
    ClassificationResponse:
      example:
        result:
          matched_policy: matched_policy
          policy_version: 0
          max_classification: null
          max_sensitivity: null
          action: flag
          handling: handling
          matched_rule: matched_rule
        record_id: record_id
        diagnostics:
          selected_policy_applied: false
          applicable_policy_count: 6
          out_of_scope_reason: none
          policy_scope_summary:
            sources:
            - sources
            - sources
            jurisdiction: jurisdiction
            excluded_data_types:
            - excluded_data_types
            - excluded_data_types
            excluded_sources:
            - excluded_sources
            - excluded_sources
            data_types:
            - data_types
            - data_types
        jurisdiction_applied: jurisdiction_applied
        classification_id: classification_id
        detected_entities:
        - snippet: snippet
          field: field
          regulatory_refs:
          - regulatory_refs
          - regulatory_refs
          confidence: definitive
          label: label
          sensitivity: low
          entity_id: entity_id
          classification: PII
          detector: detector
        - snippet: snippet
          field: field
          regulatory_refs:
          - regulatory_refs
          - regulatory_refs
          confidence: definitive
          label: label
          sensitivity: low
          entity_id: entity_id
          classification: PII
          detector: detector
        audit_ref: audit_ref
        classified_at: classified_at
      properties:
        record_id:
          title: Record Id
          type: string
        classification_id:
          title: Classification Id
          type: string
        detected_entities:
          items:
            $ref: "#/components/schemas/DetectedEntity"
          type: array
        result:
          $ref: "#/components/schemas/ClassificationResultDetail"
        diagnostics:
          $ref: "#/components/schemas/ClassificationDiagnostics"
        jurisdiction_applied:
          nullable: true
          type: string
        classified_at:
          title: Classified At
          type: string
        audit_ref:
          nullable: true
          type: string
      required:
      - classification_id
      - classified_at
      - record_id
      - result
      title: ClassificationResponse
...
    DsarRequestStatus:
      enum:
      - received
      - in_progress
      - completed
      - partial
      - denied
      - failed
      title: DsarRequestStatus
      type: string
    DsarRequestType:
      enum:
      - access
      - erasure
      title: DsarRequestType
      type: string
...
    DsarSubmitRequest:
      description: POST /api/v1/dsar/access|erasure body.
      example:
        subject_id: subject_id
        policy_id: policy_id
        records:
        - record_id: record_id
          metadata:
            key: ""
          jurisdiction: jurisdiction
          data_type: data_type
          source: source
        - record_id: record_id
          metadata:
            key: ""
          jurisdiction: jurisdiction
          data_type: data_type
          source: source
        identity:
          key: ""
      properties:
        subject_id:
          title: Subject Id
          type: string
        policy_id:
          title: Policy Id
          type: string
        identity:
          additionalProperties: true
          nullable: true
          type: object
        records:
          items:
            $ref: "#/components/schemas/RecordRef"
          nullable: true
          type: array
      required:
      - policy_id
      - subject_id
      title: DsarSubmitRequest
    EnforceRequest:
      description: POST /api/v1/enforce body.
      example:
        policy_id: policy_id
        records:
        - record_id: record_id
          metadata:
            key: ""
          jurisdiction: jurisdiction
          data_type: data_type
          source: source
        - record_id: record_id
          metadata:
            key: ""
          jurisdiction: jurisdiction
          data_type: data_type
          source: source
      properties:
        policy_id:
          nullable: true
          type: string
        records:
          items:
            $ref: "#/components/schemas/RecordRef"
          nullable: true
          type: array
      title: EnforceRequest
    EnforceResponse:
      example:
        job_id: job_id
        status: queued
      properties:
        job_id:
          title: Job Id
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
      required:
      - job_id
      - status
      title: EnforceResponse
    EnforcementJob:
      example:
        policy_id: policy_id
        finished_at: 2000-01-23T04:56:07.000+00:00
        inline_records:
        - record_id: record_id
          metadata:
            key: ""
          jurisdiction: jurisdiction
          data_type: data_type
          source: source
        - record_id: record_id
          metadata:
            key: ""
          jurisdiction: jurisdiction
          data_type: data_type
          source: source
        started_at: 2000-01-23T04:56:07.000+00:00
        progress:
          dispatched: 6
          scanned: 0
          notified: 5
          errors: 5
          pending_grace: 1
        id: id
        trigger: schedule
        error: error
        status: queued
        requested_at: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          title: Id
          type: string
        policy_id:
          nullable: true
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        trigger:
          $ref: "#/components/schemas/JobTrigger"
        requested_at:
          format: date-time
          title: Requested At
          type: string
        started_at:
          format: date-time
          nullable: true
          type: string
        finished_at:
          format: date-time
          nullable: true
          type: string
        progress:
          $ref: "#/components/schemas/JobProgress"
        error:
          nullable: true
          type: string
        inline_records:
          items:
            $ref: "#/components/schemas/RecordRef"
          nullable: true
          type: array
      required:
      - id
      - requested_at
      title: EnforcementJob
...
    EvaluationRequest:
      description: Request to evaluate a record against policies.
      example:
        record_id: record_id
        metadata:
          key: ""
        jurisdiction: jurisdiction
        data_type: data_type
        context:
          key: ""
        source: source
      properties:
        data_type:
          title: Data Type
          type: string
        record_id:
          title: Record Id
          type: string
        metadata:
          additionalProperties: true
          title: Metadata
          type: object
        source:
          nullable: true
          type: string
        jurisdiction:
          nullable: true
          type: string
        context:
          additionalProperties: true
          nullable: true
          type: object
      required:
      - data_type
      - record_id
      title: EvaluationRequest
    EvaluationResponse:
      description: Full evaluation response (API + SDK).
      example:
        result:
          matched_policy: matched_policy
          policy_version: 0
          requires_approval: false
          retain_until: retain_until
          archive_target: archive_target
          notify_at: notify_at
          confidence: none
          action: retain
          matched_rule: matched_rule
          grace_period_ends: grace_period_ends
        record_id: record_id
        evaluation_id: evaluation_id
        conflicting_policies:
        - rule_id: rule_id
          policy_id: policy_id
          action: null
          priority: 6
        - rule_id: rule_id
          policy_id: policy_id
          action: null
          priority: 6
        jurisdiction_applied: jurisdiction_applied
        audit_ref: audit_ref
        evaluated_at: evaluated_at
      properties:
        record_id:
          title: Record Id
          type: string
        evaluation_id:
          title: Evaluation Id
          type: string
        result:
          $ref: "#/components/schemas/EvaluationResultDetail"
        conflicting_policies:
          items:
            $ref: "#/components/schemas/ConflictingPolicy"
          type: array
        jurisdiction_applied:
          nullable: true
          type: string
        evaluated_at:
          title: Evaluated At
          type: string
        audit_ref:
          nullable: true
          type: string
      required:
      - evaluated_at
      - evaluation_id
      - record_id
      - result
      title: EvaluationResponse
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          input: ""
          ctx: "{}"
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          input: ""
          ctx: "{}"
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    JobStatus:
      enum:
      - queued
      - running
      - succeeded
      - failed
      - cancelled
      title: JobStatus
      type: string
    JobTrigger:
      enum:
      - schedule
      - api
      title: JobTrigger
      type: string
...
  securitySchemes:
    BearerAuth:
      description: API key as Bearer token (DRPE_API_KEY)
      scheme: bearer
      type: http